phoenix-engine/python/requirements.txt
# Phoenix AI Trading System - Python Training Service Requirements

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ONNX for model deployment
onnx>=1.15.0
onnxruntime>=1.16.0
skl2onnx>=1.15.0

# Supabase client
supabase>=2.3.0

# HTTP requests
requests>=2.31.0

# Async support
aiohttp>=3.8.0

# Data fetching
ccxt>=4.0.0  # For crypto exchange APIs

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web server for deployment
fastapi>=0.103.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
structlog>=23.1.0

# Existing dependencies
pyzmq>=25.1.1
pycapnp>=2.0.2

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
